requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-generativeai>=0.8.0
requests==2.*
python-dotenv==1.*